Reworking the framework from .js to .ts
