+++ cypress/e2e/products.cy.ts
@@ -1,0 +1,16 @@
+import Home from '../src/pages/Home';
+
+const homePage = new Home();
+
+describe('Check product on site', () => {
+  it('Check item discount', () => {
+    homePage
+      .visit()
+      .checkPageUrl()
+      .productList
+      .clickOnBestSellersBtn()
+      .openQuickViewModalForFirstProductWithPriceDiscount()
+      .waitForLoad()
+      .checkSumWithDiscountCalculatedCorrectly();
+  });
+});
