+++ cypress/src/pages/Account.ts
@@ -2,18 +2,18 @@ import Header from '../components/Header';
 import BasePage from '../base/BasePage';
 
 export default class Account extends BasePage {
-  header = new Header();
-  accountInfoContainerLocator = '#center_column';
+  header: Header = new Header();
+  private accountInfoContainerLocator: string = '#center_column';
 
   constructor() {
     super('index.php?controller=my-account');
   }
 
-  get accountInfo() {
+  private get accountInfo(): Cypress.Chainable<JQuery> {
     return cy.get(this.accountInfoContainerLocator);
   }
 
-  checkTextIsPresent(expectedText) {
+  public checkTextIsPresent(expectedText: string): this {
     this.accountInfo.contains(expectedText);
     return this;
   }
