--- cypress/src/pages/ContactUs.js
@@ -1,7 +1,0 @@
-import BasePage from '../base/BasePage';
-
-export default class ContactUs extends BasePage{
-    constructor() {
-        super('index.php?controller=contact');
-    }
-}
