+++ cypress/src/pages/ContactUs.ts
@@ -1,0 +1,7 @@
+import BasePage from '../base/BasePage';
+
+export default class ContactUs extends BasePage {
+  constructor() {
+    super('index.php?controller=contact');
+  }
+}
