+++ cypress/src/pages/Home.ts
@@ -4,17 +4,16 @@ import Footer from '../components/Footer';
 import ProductList from '../components/ProductList';
 
 export default class Home extends BasePage {
-    header = new Header();
-    footer = new Footer();
-    productList = new ProductList();
+  header: Header = new Header();
+  footer: Footer = new Footer();
+  productList: ProductList = new ProductList();
 
+  constructor() {
+    super('index.php');
+  }
 
-    constructor() {
-        super('index.php');
-    }
-
-    visit() {
-        cy.visit('/');
-        return this;
-    }
+  public visit(): this {
+    cy.visit('/');
+    return this;
+  }
 }
