+++ package.json
@@ -19,10 +19,12 @@
   },
   "homepage": "https://github.com/VadimNastoyashchy/cypressautomationpracticecom#readme",
   "devDependencies": {
-    "cypress": "10.2.0",
+    "@types/node": "^20.11.16",
+    "cypress": "^10.2.0",
     "cypress-iframe": "1.0.1",
     "eslint": "^8.56.0",
     "eslint-plugin-cypress": "^2.12.1",
-    "mochawesome": "7.1.3"
+    "mochawesome": "7.1.3",
+    "typescript": "^5.3.3"
   }
 }
